adding launch options for puppeteer
